requirements-prod.txt

# Async HTTP Client
aiohttp>=3.8.0
aiosqlite>=0.17.0
alembic>=1.7.3

# Development
black>=21.7b0
# Core Dependencies
fastapi>=0.68.0
flake8>=3.9.2

# API Documentation
gunicorn>=20.1.0
httpx>=0.19.0
isort>=5.9.3

# Documentation
mkdocs>=1.2.3
mkdocs-material>=7.3.6
mypy>=0.910
passlib[bcrypt]>=1.7.4

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
python-dotenv>=0.19.0
python-jose[cryptography]>=3.3.0
python-jose[cryptography]>=3.3.0  # For JWT tokens

# Security
python-keycloak>=2.8.0  # For OAuth2/OpenID Connect
python-multipart>=0.0.5

# Configuration
pyyaml>=5.4.1

# Database
sqlalchemy>=1.4.23
uvicorn>=0.15.0
uvloop>=0.16.0
